+++ src/main/java/com/example/fashionforecastbackend/customOutfit/dto/request/MemberOutfitRequest.java
@@ -1,4 +1,4 @@
-package com.example.fashionforecastbackend.member.dto.request;
+package com.example.fashionforecastbackend.customOutfit.dto.request;
 
 import jakarta.validation.constraints.NotNull;
