+++ src/main/java/com/example/fashionforecastbackend/member/domain/repository/MemberOutfitRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-import com.example.fashionforecastbackend.member.domain.MemberOutfit;
+import com.example.fashionforecastbackend.customOutfit.domain.MemberOutfit;
 
 public interface MemberOutfitRepository extends JpaRepository<MemberOutfit, Long> {
